Refactor: extract controllers and simplify SurveyController and RespondentController (#2050)
* Refactor: extract QuestionnaireSimulationController

Extracts the simulation methods out of QuestionnaireController in
order to have a distinct controller to run a questionnaire
simulation, focusing the original QuestionnaireController to CRUD
operations over questionnaires.

* Refactor: extract SurveyQuestionnaireController

Extracts the simulation methods out of SurveyController in order to
have a distinct controller to simulate a questionaire across a real
channel using a temporary one-respondent simulation survey.

* Refactor: extract SurveyLinkController

Extracts the link methods out of SurveyController in order to have a
distinct controller to create, refresh and delete survey links.

Now SurveyController is focused on CRUD operations on Surveys as
well as launching and stopping survey runs.

* Add load_survey method to SurveyController and RespondentController

Simplifies the controllers by extracting repeated code in every
function, and simplify a little bit the logic of a few other methods
along the way.
