+++ test/controllers/survey_controller_test.exs
@@ -612,199 +612,6 @@ defmodule Ask.SurveyControllerTest do
     end
   end
 
-  describe "stop simulation" do
-    setup %{conn: conn, user: user} do
-      create_running_survey = fn simulation ->
-        [survey | _tail] =
-          create_running_survey_with_channel_and_respondent_with_options(
-            user: user,
-            simulation: simulation
-          )
-
-        %{survey: survey}
-      end
-
-      stop_simulation = fn survey ->
-        post(
-          conn,
-          project_survey_survey_path(
-            conn,
-            :stop_simulation,
-            survey.project,
-            survey
-          )
-        )
-      end
-
-      {:ok, create_running_survey: create_running_survey, stop_simulation: stop_simulation}
-    end
-
-    test "limits endpoint for survey simulations only", %{
-      create_running_survey: create_running_survey,
-      stop_simulation: stop_simulation
-    } do
-      simulation = false
-      %{survey: survey} = create_running_survey.(simulation)
-
-      assert_error_sent(:not_found, fn ->
-        stop_simulation.(survey)
-      end)
-    end
-  end
-
-  describe "simulation status" do
-    setup %{conn: conn, user: user} do
-      create_running_survey = fn simulation ->
-        [survey | _tail] =
-          create_running_survey_with_channel_and_respondent_with_options(
-            user: user,
-            simulation: simulation
-          )
-
-        %{survey: survey}
-      end
-
-      get_simulation_status = fn survey ->
-        get(
-          conn,
-          project_survey_survey_path(
-            conn,
-            :simulation_status,
-            survey.project,
-            survey
-          )
-        )
-      end
-
-      {:ok,
-       create_running_survey: create_running_survey, get_simulation_status: get_simulation_status}
-    end
-
-    test "limits endpoint for survey simulations only", %{
-      create_running_survey: create_running_survey,
-      get_simulation_status: get_simulation_status
-    } do
-      simulation = false
-      %{survey: survey} = create_running_survey.(simulation)
-
-      assert_error_sent(:not_found, fn ->
-        get_simulation_status.(survey)
-      end)
-    end
-  end
-
-  describe "simulation initial state" do
-    setup %{conn: conn, user: user} do
-      create_running_survey = fn mode, simulation ->
-        [survey, _group, _test_channel, respondent, _phone_number] =
-          create_running_survey_with_channel_and_respondent_with_options(
-            user: user,
-            mode: mode,
-            simulation: simulation
-          )
-
-        %{survey: survey, respondent_id: respondent.id}
-      end
-
-      get_simulation_initial_state = fn survey, mode ->
-        get(
-          conn,
-          project_survey_survey_path(
-            conn,
-            :simulation_initial_state,
-            survey.project,
-            survey,
-            mode
-          )
-        )
-      end
-
-      poll_survey = fn ->
-        Broker.start_link()
-        {:ok, _pid} = ChannelStatusServer.start_link()
-        Process.register(self(), :mail_target)
-        Broker.poll()
-      end
-
-      mobile_contact_messages = fn respondent_id ->
-        Repo.get!(Respondent, respondent_id)
-        |> Session.load_respondent_session(true)
-        |> Session.mobile_contact_message()
-      end
-
-      {:ok,
-       create_running_survey: create_running_survey,
-       get_simulation_initial_state: get_simulation_initial_state,
-       poll_survey: poll_survey,
-       mobile_contact_messages: mobile_contact_messages}
-    end
-
-    test "limits endpoint for survey simulations only", %{
-      create_running_survey: create_running_survey,
-      get_simulation_initial_state: get_simulation_initial_state
-    } do
-      {mode, simulation} = {"sms", false}
-      %{survey: survey} = create_running_survey.(mode, simulation)
-
-      assert_error_sent(:not_found, fn ->
-        get_simulation_initial_state.(survey, mode)
-      end)
-    end
-
-    test "SMS return an empty map", %{
-      create_running_survey: create_running_survey,
-      get_simulation_initial_state: get_simulation_initial_state
-    } do
-      {mode, simulation} = {"sms", true}
-      %{survey: survey} = create_running_survey.(mode, simulation)
-
-      conn = get_simulation_initial_state.(survey, mode)
-
-      assert json_response(conn, 200)["data"] == %{}
-    end
-
-    test "IVR return an empty map", %{
-      create_running_survey: create_running_survey,
-      get_simulation_initial_state: get_simulation_initial_state
-    } do
-      {mode, simulation} = {"ivr", true}
-      %{survey: survey} = create_running_survey.(mode, simulation)
-
-      conn = get_simulation_initial_state.(survey, mode)
-
-      assert json_response(conn, 200)["data"] == %{}
-    end
-
-    test "Mobileweb fails when the respondent isn't ready", %{
-      create_running_survey: create_running_survey,
-      get_simulation_initial_state: get_simulation_initial_state
-    } do
-      {mode, simulation} = {"mobileweb", true}
-      %{survey: survey} = create_running_survey.(mode, simulation)
-
-      conn = get_simulation_initial_state.(survey, mode)
-
-      %{status: status} = conn
-      assert status == 404
-    end
-
-    test "Mobileweb answers when the respondent is ready", %{
-      create_running_survey: create_running_survey,
-      mobile_contact_messages: mobile_contact_messages,
-      poll_survey: poll_survey,
-      get_simulation_initial_state: get_simulation_initial_state
-    } do
-      {mode, simulation} = {"mobileweb", true}
-      %{survey: survey, respondent_id: respondent_id} = create_running_survey.(mode, simulation)
-      poll_survey.()
-
-      conn = get_simulation_initial_state.(survey, mode)
-
-      assert json_response(conn, 200)["data"]["mobile_contact_messages"] ==
-               mobile_contact_messages.(respondent_id)
-    end
-  end
-
   describe "count_partial_results stats" do
     setup %{conn: conn, user: user} do
       survey = count_partial_results_test_survey(%{
@@ -2077,335 +1884,19 @@ defmodule Ask.SurveyControllerTest do
     end
   end
 
-  describe "download links" do
-    test "results link generation", %{conn: conn, user: user} do
-      project = create_project_for_user(user)
+  describe "config" do
+    setup context do
+      project = create_project_for_user(context[:user])
       survey = insert(:survey, project: project)
-
-      response = get conn, project_survey_links_path(conn, :create_link, project, survey, "results")
-
-      link = ShortLink |> Repo.one
-
-      assert json_response(response, 200) == %{
-        "name" => "survey/#{survey.id}/results",
-        "url" => "#{Ask.Endpoint.url}/link/#{link.hash}"
-      }
-
-      assert link.target == "/api/v1/projects/#{project.id}/surveys/#{survey.id}/respondents/results?_format=csv"
-
-      response = get conn, project_survey_links_path(conn, :create_link, project, survey, "results")
-
-      assert json_response(response, 200) == %{
-        "name" => "survey/#{survey.id}/results",
-        "url" => "#{Ask.Endpoint.url}/link/#{link.hash}"
-      }
-      assert ShortLink |> Repo.all |> length == 1
-
-      response = put conn, project_survey_links_path(conn, :refresh_link, project, survey, "results")
-
-      new_link = ShortLink |> Repo.one
-
-      assert json_response(response, 200) == %{
-        "name" => "survey/#{survey.id}/results",
-        "url" => "#{Ask.Endpoint.url}/link/#{new_link.hash}"
-      }
-
-      assert link.hash != new_link.hash
-      assert link.target == new_link.target
-
-      response = delete conn, project_survey_links_path(conn, :delete_link, project, survey, "results")
-
-      assert response(response, 204)
-      assert [] == ShortLink |> Repo.all()
+      [project: project, survey: survey]
     end
 
-    test "incentives link generation", %{conn: conn, user: user} do
-      project = create_project_for_user(user)
-      survey = insert(:survey, project: project)
-
-      response = get conn, project_survey_links_path(conn, :create_link, project, survey, "incentives")
-
-      link = ShortLink |> Repo.one
-
-      assert json_response(response, 200) == %{
-        "name" => "survey/#{survey.id}/incentives",
-        "url" => "#{Ask.Endpoint.url}/link/#{link.hash}"
-      }
-
-      response = get conn, project_survey_links_path(conn, :create_link, project, survey, "incentives")
-
-      assert json_response(response, 200) == %{
-        "name" => "survey/#{survey.id}/incentives",
-        "url" => "#{Ask.Endpoint.url}/link/#{link.hash}"
-      }
-      assert ShortLink |> Repo.all |> length == 1
-
-      assert link.target == "/api/v1/projects/#{project.id}/surveys/#{survey.id}/respondents/incentives?_format=csv"
-
-      response = put conn, project_survey_links_path(conn, :refresh_link, project, survey, "incentives")
-
-      new_link = ShortLink |> Repo.one
-
-      assert json_response(response, 200) == %{
-        "name" => "survey/#{survey.id}/incentives",
-        "url" => "#{Ask.Endpoint.url}/link/#{new_link.hash}"
-      }
-
-      assert link.hash != new_link.hash
-      assert link.target == new_link.target
-
-      response = delete conn, project_survey_links_path(conn, :delete_link, project, survey, "incentives")
-
-      assert response(response, 204)
-      assert [] == ShortLink |> Repo.all()
+    test "post", %{conn: conn, project: project, survey: survey} do
+      post conn, project_survey_survey_path(conn, :config, project, survey)
     end
 
-    test "interactions link generation", %{conn: conn, user: user} do
-      project = create_project_for_user(user)
-      survey = insert(:survey, project: project)
-
-      response = get conn, project_survey_links_path(conn, :create_link, project, survey, "interactions")
-
-      link = ShortLink |> Repo.one
-
-      assert json_response(response, 200) == %{
-        "name" => "survey/#{survey.id}/interactions",
-        "url" => "#{Ask.Endpoint.url}/link/#{link.hash}"
-      }
-
-      response = get conn, project_survey_links_path(conn, :create_link, project, survey, "interactions")
-
-      assert json_response(response, 200) == %{
-        "name" => "survey/#{survey.id}/interactions",
-        "url" => "#{Ask.Endpoint.url}/link/#{link.hash}"
-      }
-      assert ShortLink |> Repo.all |> length == 1
-
-      assert link.target == "/api/v1/projects/#{project.id}/surveys/#{survey.id}/respondents/interactions?_format=csv"
-
-      response = put conn, project_survey_links_path(conn, :refresh_link, project, survey, "interactions")
-
-      new_link = ShortLink |> Repo.one
-
-      assert json_response(response, 200) == %{
-        "name" => "survey/#{survey.id}/interactions",
-        "url" => "#{Ask.Endpoint.url}/link/#{new_link.hash}"
-      }
-
-      assert link.hash != new_link.hash
-      assert link.target == new_link.target
-
-      response = delete conn, project_survey_links_path(conn, :delete_link, project, survey, "interactions")
-
-      assert response(response, 204)
-      assert [] == ShortLink |> Repo.all()
-    end
-
-    test "disposition_history link generation", %{conn: conn, user: user} do
-      project = create_project_for_user(user)
-      survey = insert(:survey, project: project)
-
-      response = get conn, project_survey_links_path(conn, :create_link, project, survey, "disposition_history")
-
-      link = ShortLink |> Repo.one
-
-      assert json_response(response, 200) == %{
-        "name" => "survey/#{survey.id}/disposition_history",
-        "url" => "#{Ask.Endpoint.url}/link/#{link.hash}"
-      }
-      assert link.target == "/api/v1/projects/#{project.id}/surveys/#{survey.id}/respondents/disposition_history?_format=csv"
-
-      response = get conn, project_survey_links_path(conn, :create_link, project, survey, "disposition_history")
-
-      assert json_response(response, 200) == %{
-        "name" => "survey/#{survey.id}/disposition_history",
-        "url" => "#{Ask.Endpoint.url}/link/#{link.hash}"
-      }
-      assert ShortLink |> Repo.all |> length == 1
-
-      response = put conn, project_survey_links_path(conn, :refresh_link, project, survey, "disposition_history")
-
-      new_link = ShortLink |> Repo.one
-
-      assert json_response(response, 200) == %{
-        "name" => "survey/#{survey.id}/disposition_history",
-        "url" => "#{Ask.Endpoint.url}/link/#{new_link.hash}"
-      }
-
-      assert link.hash != new_link.hash
-      assert link.target == new_link.target
-
-      response = delete conn, project_survey_links_path(conn, :delete_link, project, survey, "disposition_history")
-
-      assert response(response, 204)
-      assert [] == ShortLink |> Repo.all()
-    end
-
-    test "forbids readers to create links", %{conn: conn, user: user} do
-      project = create_project_for_user(user, level: "reader")
-      survey = insert(:survey, project: project)
-
-      assert_error_sent :forbidden, fn ->
-        get conn, project_survey_links_path(conn, :create_link, project, survey, "results")
-      end
-      assert_error_sent :forbidden, fn ->
-        get conn, project_survey_links_path(conn, :create_link, project, survey, "incentives")
-      end
-      assert_error_sent :forbidden, fn ->
-        get conn, project_survey_links_path(conn, :create_link, project, survey, "interactions")
-      end
-      assert_error_sent :forbidden, fn ->
-        get conn, project_survey_links_path(conn, :create_link, project, survey, "disposition_history")
-      end
-    end
-
-    test "forbids readers to refresh links", %{conn: conn, user: user} do
-      project = create_project_for_user(user, level: "reader")
-      survey = insert(:survey, project: project)
-
-      assert_error_sent :forbidden, fn ->
-        put conn, project_survey_links_path(conn, :refresh_link, project, survey, "results")
-      end
-      assert_error_sent :forbidden, fn ->
-        put conn, project_survey_links_path(conn, :refresh_link, project, survey, "incentives")
-      end
-      assert_error_sent :forbidden, fn ->
-        put conn, project_survey_links_path(conn, :refresh_link, project, survey, "interactions")
-      end
-      assert_error_sent :forbidden, fn ->
-        put conn, project_survey_links_path(conn, :refresh_link, project, survey, "disposition_history")
-      end
-    end
-
-    test "forbids readers to delete links", %{conn: conn, user: user} do
-      project = create_project_for_user(user, level: "reader")
-      survey = insert(:survey, project: project)
-
-      assert_error_sent :forbidden, fn ->
-        delete conn, project_survey_links_path(conn, :delete_link, project, survey, "results")
-      end
-      assert_error_sent :forbidden, fn ->
-        delete conn, project_survey_links_path(conn, :delete_link, project, survey, "incentives")
-      end
-      assert_error_sent :forbidden, fn ->
-        delete conn, project_survey_links_path(conn, :delete_link, project, survey, "interactions")
-      end
-      assert_error_sent :forbidden, fn ->
-        delete conn, project_survey_links_path(conn, :delete_link, project, survey, "disposition_history")
-      end
-    end
-
-    test "forbids to create links if project is archived", %{conn: conn, user: user} do
-      project = create_project_for_user(user, archived: true)
-      survey = insert(:survey, project: project)
-
-      assert_error_sent :forbidden, fn ->
-        get conn, project_survey_links_path(conn, :create_link, project, survey, "results")
-      end
-      assert_error_sent :forbidden, fn ->
-        get conn, project_survey_links_path(conn, :create_link, project, survey, "incentives")
-      end
-      assert_error_sent :forbidden, fn ->
-        get conn, project_survey_links_path(conn, :create_link, project, survey, "interactions")
-      end
-      assert_error_sent :forbidden, fn ->
-        get conn, project_survey_links_path(conn, :create_link, project, survey, "disposition_history")
-      end
-    end
-
-    test "forbids to refresh links if project is archived", %{conn: conn, user: user} do
-      project = create_project_for_user(user, archived: true)
-      survey = insert(:survey, project: project)
-
-      assert_error_sent :forbidden, fn ->
-        put conn, project_survey_links_path(conn, :refresh_link, project, survey, "results")
-      end
-      assert_error_sent :forbidden, fn ->
-        put conn, project_survey_links_path(conn, :refresh_link, project, survey, "incentives")
-      end
-      assert_error_sent :forbidden, fn ->
-        put conn, project_survey_links_path(conn, :refresh_link, project, survey, "interactions")
-      end
-      assert_error_sent :forbidden, fn ->
-        put conn, project_survey_links_path(conn, :refresh_link, project, survey, "disposition_history")
-      end
-    end
-
-    test "forbids to delete links if project is archived", %{conn: conn, user: user} do
-      project = create_project_for_user(user, archived: true)
-      survey = insert(:survey, project: project)
-
-      assert_error_sent :forbidden, fn ->
-        delete conn, project_survey_links_path(conn, :delete_link, project, survey, "results")
-      end
-      assert_error_sent :forbidden, fn ->
-        delete conn, project_survey_links_path(conn, :delete_link, project, survey, "incentives")
-      end
-      assert_error_sent :forbidden, fn ->
-        delete conn, project_survey_links_path(conn, :delete_link, project, survey, "interactions")
-      end
-      assert_error_sent :forbidden, fn ->
-        delete conn, project_survey_links_path(conn, :delete_link, project, survey, "disposition_history")
-      end
-    end
-
-    test "allows editors to create some links", %{conn: conn, user: user} do
-      project = create_project_for_user(user, level: "editor")
-      survey = insert(:survey, project: project)
-
-      response = get conn, project_survey_links_path(conn, :create_link, project, survey, "results")
-      assert response(response, 200)
-
-      response = get conn, project_survey_links_path(conn, :create_link, project, survey, "disposition_history")
-      assert response(response, 200)
-
-      assert_error_sent :forbidden, fn ->
-        get conn, project_survey_links_path(conn, :create_link, project, survey, "incentives")
-      end
-
-      assert_error_sent :forbidden, fn ->
-        get conn, project_survey_links_path(conn, :create_link, project, survey, "interactions")
-      end
-    end
-
-    test "allows editors to refresh some links", %{conn: conn, user: user} do
-      project = create_project_for_user(user, level: "editor")
-      survey = insert(:survey, project: project)
-      get conn, project_survey_links_path(conn, :create_link, project, survey, "results")
-      get conn, project_survey_links_path(conn, :create_link, project, survey, "disposition_history")
-
-      response = put conn, project_survey_links_path(conn, :refresh_link, project, survey, "results")
-      assert response(response, 200)
-
-      response = put conn, project_survey_links_path(conn, :refresh_link, project, survey, "disposition_history")
-      assert response(response, 200)
-
-      assert_error_sent :forbidden, fn ->
-        put conn, project_survey_links_path(conn, :refresh_link, project, survey, "incentives")
-      end
-      assert_error_sent :forbidden, fn ->
-        put conn, project_survey_links_path(conn, :refresh_link, project, survey, "interactions")
-      end
-    end
-
-    test "forbids editor to delete some links", %{conn: conn, user: user} do
-      project = create_project_for_user(user, level: "editor")
-      survey = insert(:survey, project: project)
-      get conn, project_survey_links_path(conn, :create_link, project, survey, "results")
-      get conn, project_survey_links_path(conn, :create_link, project, survey, "disposition_history")
-
-      response = delete conn, project_survey_links_path(conn, :delete_link, project, survey, "results")
-      assert response(response, 204)
-
-      response = delete conn, project_survey_links_path(conn, :delete_link, project, survey, "disposition_history")
-      assert response(response, 204)
-
-      assert_error_sent :forbidden, fn ->
-        delete conn, project_survey_links_path(conn, :delete_link, project, survey, "incentives")
-      end
-      assert_error_sent :forbidden, fn ->
-        delete conn, project_survey_links_path(conn, :delete_link, project, survey, "interactions")
-      end
+    test "get", %{conn: conn, project: project, survey: survey} do
+      get conn, project_survey_survey_path(conn, :config, project, survey)
     end
   end
 
@@ -2414,74 +1905,6 @@ defmodule Ask.SurveyControllerTest do
       remote_ip = {192, 168, 0, 128}
       conn = %{conn | remote_ip: remote_ip}
       {:ok, conn: conn, remote_ip: remote_ip}
-    end
-
-    test "generates logs for results link", %{conn: conn, user: user} do
-      project = create_project_for_user(user)
-      survey = insert(:survey, project: project)
-
-      get conn, project_survey_links_path(conn, :create_link, project, survey, "results")
-      activity_log_create = ActivityLog |> where([log], log.action == "enable_public_link") |> Repo.one
-      assert_link_log(%{log: activity_log_create, user: user, project: project, survey: survey, action: "enable_public_link", report_type: "survey_results", remote_ip: "192.168.0.128"})
-
-      put conn, project_survey_links_path(conn, :refresh_link, project, survey, "results")
-      activity_log_regenerate = ActivityLog |> where([log], log.action == "regenerate_public_link") |> Repo.one
-      assert_link_log(%{log: activity_log_regenerate, user: user, project: project, survey: survey, action: "regenerate_public_link", report_type: "survey_results", remote_ip: "192.168.0.128"})
-
-      delete conn, project_survey_links_path(conn, :delete_link, project, survey, "results")
-      activity_log_delete = ActivityLog |> where([log], log.action == "disable_public_link") |> Repo.one
-      assert_link_log(%{log: activity_log_delete, user: user, project: project, survey: survey, action: "disable_public_link", report_type: "survey_results", remote_ip: "192.168.0.128"})
-    end
-
-    test "generates logs for incentives link", %{conn: conn, user: user} do
-      project = create_project_for_user(user)
-      survey = insert(:survey, project: project)
-
-      get conn, project_survey_links_path(conn, :create_link, project, survey, "incentives")
-      activity_log_create = ActivityLog |> where([log], log.action == "enable_public_link") |> Repo.one
-      assert_link_log(%{log: activity_log_create, user: user, project: project, survey: survey, action: "enable_public_link", report_type: "incentives", remote_ip: "192.168.0.128"})
-
-      put conn, project_survey_links_path(conn, :refresh_link, project, survey, "incentives")
-      activity_log_regenerate = ActivityLog |> where([log], log.action == "regenerate_public_link") |> Repo.one
-      assert_link_log(%{log: activity_log_regenerate, user: user, project: project, survey: survey, action: "regenerate_public_link", report_type: "incentives", remote_ip: "192.168.0.128"})
-
-      delete conn, project_survey_links_path(conn, :delete_link, project, survey, "incentives")
-      activity_log_delete = ActivityLog |> where([log], log.action == "disable_public_link") |> Repo.one
-      assert_link_log(%{log: activity_log_delete, user: user, project: project, survey: survey, action: "disable_public_link", report_type: "incentives", remote_ip: "192.168.0.128"})
-    end
-
-    test "generates logs for interactions link", %{conn: conn, user: user} do
-      project = create_project_for_user(user)
-      survey = insert(:survey, project: project)
-
-      get conn, project_survey_links_path(conn, :create_link, project, survey, "interactions")
-      activity_log_create = ActivityLog |> where([log], log.action == "enable_public_link") |> Repo.one
-      assert_link_log(%{log: activity_log_create, user: user, project: project, survey: survey, action: "enable_public_link", report_type: "interactions", remote_ip: "192.168.0.128"})
-
-      put conn, project_survey_links_path(conn, :refresh_link, project, survey, "interactions")
-      activity_log_regenerate = ActivityLog |> where([log], log.action == "regenerate_public_link") |> Repo.one
-      assert_link_log(%{log: activity_log_regenerate, user: user, project: project, survey: survey, action: "regenerate_public_link", report_type: "interactions", remote_ip: "192.168.0.128"})
-
-      delete conn, project_survey_links_path(conn, :delete_link, project, survey, "interactions")
-      activity_log_delete = ActivityLog |> where([log], log.action == "disable_public_link") |> Repo.one
-      assert_link_log(%{log: activity_log_delete, user: user, project: project, survey: survey, action: "disable_public_link", report_type: "interactions", remote_ip: "192.168.0.128"})
-    end
-
-    test "generates logs for disposition_history link", %{conn: conn, user: user} do
-      project = create_project_for_user(user)
-      survey = insert(:survey, project: project)
-
-      get conn, project_survey_links_path(conn, :create_link, project, survey, "disposition_history")
-      activity_log_create = ActivityLog |> where([log], log.action == "enable_public_link") |> Repo.one
-      assert_link_log(%{log: activity_log_create, user: user, project: project, survey: survey, action: "enable_public_link", report_type: "disposition_history", remote_ip: "192.168.0.128"})
-
-      put conn, project_survey_links_path(conn, :refresh_link, project, survey, "disposition_history")
-      activity_log_regenerate = ActivityLog |> where([log], log.action == "regenerate_public_link") |> Repo.one
-      assert_link_log(%{log: activity_log_regenerate, user: user, project: project, survey: survey, action: "regenerate_public_link", report_type: "disposition_history", remote_ip: "192.168.0.128"})
-
-      delete conn, project_survey_links_path(conn, :delete_link, project, survey, "disposition_history")
-      activity_log_delete = ActivityLog |> where([log], log.action == "disable_public_link") |> Repo.one
-      assert_link_log(%{log: activity_log_delete, user: user, project: project, survey: survey, action: "disable_public_link", report_type: "disposition_history", remote_ip: "192.168.0.128"})
     end
 
     test "generates log after creating a survey", %{conn: conn, user: user} do
@@ -2665,14 +2088,6 @@ defmodule Ask.SurveyControllerTest do
     assert log.entity_type == "survey"
     assert log.action == action
     assert log.remote_ip == remote_ip
-  end
-
-  defp assert_link_log(%{log: log, user: user, project: project, survey: survey, action: action, report_type: report_type, remote_ip: remote_ip}) do
-    assert_log(log, user.id, project, survey, action, remote_ip)
-    assert log.metadata == %{
-      "survey_name" => survey.name,
-      "report_type" => report_type
-    }
   end
 
   defp assert_survey_log(%{log: log, user: user, project: project, survey: survey, action: action, remote_ip: remote_ip, metadata: metadata}) do
