+++ web/controllers/questionnaire_controller.ex
@@ -15,11 +15,7 @@ defmodule Ask.QuestionnaireController do
     Gettext
   }
   alias Ecto.Multi
-  alias Ask.Runtime.{
-    QuestionnaireSimulator,
-    QuestionnaireMobileWebSimulator,
-    QuestionnaireAction
-  }
+  alias Ask.Runtime.QuestionnaireAction
 
   plug :validate_params when action in [:create, :update]
   action_fallback Ask.FallbackController
@@ -336,30 +332,6 @@ defmodule Ask.QuestionnaireController do
     |> Repo.update!
 
     render(conn, "show.json", questionnaire: questionnaire)
-  end
-
-  def start_simulation(conn, %{"project_id" => project_id, "questionnaire_id" => id}) do
-    project = conn |> load_project(project_id)
-    mode = conn.params["mode"]
-    with {:ok, questionnaire} <- load_questionnaire(project, id),
-         {:ok, simulation_response} <- QuestionnaireSimulator.start_simulation(project, questionnaire, mode)
-    do
-      render(conn, "simulation.json", simulation: simulation_response, mode: mode)
-    end
-  end
-
-  def sync_simulation(conn, %{"project_id" => project_id, "respondent_id" => respondent_id, "response" => response, "mode" => mode}) do
-    # Load project to authorize connection
-    conn |> load_project(project_id)
-    with {:ok, simulation_response} <- QuestionnaireSimulator.process_respondent_response(respondent_id, response, mode), do:
-      render(conn, "simulation.json", simulation: simulation_response)
-  end
-
-  def get_last_simulation_response(conn, %{"project_id" => project_id, "respondent_id" => respondent_id}) do
-    # Load project to authorize connection
-    conn |> load_project(project_id)
-    with {:ok, simulation_response} <- QuestionnaireMobileWebSimulator.get_last_simulation_response(respondent_id), do:
-      render(conn, "simulation.json", simulation: simulation_response)
   end
 
   defp validate_params(conn, _params) do
