+++ web/controllers/survey_controller.ex
@@ -1,13 +1,12 @@
 defmodule Ask.SurveyController do
   use Ask.Web, :api_controller
 
-  alias Ask.{Project, Folder, Survey, Questionnaire, Logger, RespondentGroup, Respondent, Channel, ShortLink, ActivityLog, RetriesHistogram, ScheduleError, ConflictError}
-  alias Ask.Runtime.{Session, SurveyAction}
+  alias Ask.{Project, Folder, Survey, Logger, ActivityLog, RetriesHistogram, ScheduleError, ConflictError}
+  alias Ask.Runtime.SurveyAction
   alias Ecto.Multi
 
   def index(conn, %{"project_id" => project_id} = params) do
-    project = conn
-    |> load_project(project_id)
+    project = load_project(conn, project_id)
 
     dynamic = dynamic([s], s.project_id == ^project.id and is_nil(s.folder_id) and is_nil(s.panel_survey_id))
 
@@ -42,11 +41,7 @@ defmodule Ask.SurveyController do
   end
 
   def create(conn, params = %{"project_id" => project_id}) do
-    folder_id = Map.get(params, "folder_id")
-
-    project = conn
-    |> load_project_for_change(project_id)
-    |> validate_project_not_archived(conn)
+    project = load_project_for_change(conn, project_id)
 
     survey_params = Map.get(params, "survey", %{})
     timezone = Map.get(survey_params, "timezone", Ask.Schedule.default_timezone())
@@ -54,7 +49,7 @@ defmodule Ask.SurveyController do
     generates_panel_survey = Map.get(survey_params, "generates_panel_survey", false)
     props = %{
       "project_id" => project_id,
-      "folder_id" => folder_id,
+      "folder_id" => Map.get(params, "folder_id"),
       "name" => "",
       "schedule" => schedule,
       "generates_panel_survey" => generates_panel_survey
@@ -96,8 +91,7 @@ defmodule Ask.SurveyController do
   def show(conn, %{"project_id" => project_id, "id" => id}) do
     survey = conn
     |> load_project(project_id)
-    |> assoc(:surveys)
-    |> Repo.get!(id)
+    |> load_survey(id)
     |> Repo.preload([:quota_buckets])
     |> Repo.preload(:questionnaires)
     |> Repo.preload(:folder)
@@ -110,34 +104,26 @@ defmodule Ask.SurveyController do
   end
 
   def stats(conn, %{"project_id" => project_id, "survey_id" => survey_id}) do
-    survey = conn
+    stats = conn
     |> load_project(project_id)
-    |> assoc(:surveys)
-    |> Repo.get!(survey_id)
-
-    stats = survey |> Survey.stats
+    |> load_survey(survey_id)
+    |> Survey.stats()
 
     render(conn, "stats.json", stats)
   end
 
   def retries_histograms(conn, %{"project_id" => project_id, "survey_id" => survey_id}) do
-    survey = conn
+    retries_histograms = conn
     |> load_project(project_id)
-    |> assoc(:surveys)
-    |> Repo.get!(survey_id)
-
-    retries_histograms = survey |> RetriesHistogram.survey_histograms()
+    |> load_survey(survey_id)
+    |> RetriesHistogram.survey_histograms()
 
     render(conn, "retries_histograms.json", %{histograms: retries_histograms})
   end
 
   def update(conn, %{"project_id" => project_id, "id" => id, "survey" => survey_params}) do
-    project = conn
-      |> load_project_for_change(project_id)
-
-    survey = project
-      |> assoc(:surveys)
-      |> Repo.get!(id)
+    project = load_project_for_change(conn, project_id)
+    survey = load_survey(project, id)
 
     if survey |> Survey.editable? do
       changeset = survey
@@ -184,14 +170,8 @@ defmodule Ask.SurveyController do
   end
 
   def set_folder_id(conn, %{"project_id" => project_id, "survey_id" => survey_id, "folder_id" => folder_id}) do
-    project =
-      conn
-      |> load_project_for_change(project_id)
-
-    survey =
-      project
-      |> assoc(:surveys)
-      |> Repo.get!(survey_id)
+    project = load_project_for_change(conn, project_id)
+    survey = load_survey(project, survey_id)
 
     # Panel surveys can belong to a folder, but their waves don't.
     if Survey.belongs_to_panel_survey?(survey), do: raise ConflictError
@@ -220,14 +200,8 @@ defmodule Ask.SurveyController do
 
 
   def set_name(conn, %{"project_id" => project_id, "survey_id" => survey_id, "name" => name}) do
-    project =
-      conn
-      |> load_project_for_change(project_id)
-
-    survey =
-      project
-      |> assoc(:surveys)
-      |> Repo.get!(survey_id)
+    project = load_project_for_change(conn, project_id)
+    survey = load_survey(project, survey_id)
 
     result =
       Multi.new()
@@ -248,14 +222,8 @@ defmodule Ask.SurveyController do
   end
 
   def set_description(conn, %{"project_id" => project_id, "survey_id" => survey_id, "description" => description}) do
-    project =
-      conn
-      |> load_project_for_change(project_id)
-
-    survey =
-      project
-      |> assoc(:surveys)
-      |> Repo.get!(survey_id)
+    project = load_project_for_change(conn, project_id)
+    survey = load_survey(project, survey_id)
 
     result =
       Multi.new()
@@ -284,12 +252,8 @@ defmodule Ask.SurveyController do
   end
 
   def delete(conn, %{"project_id" => project_id, "id" => id}) do
-    project = conn
-    |> load_project_for_change(project_id)
-
-    survey = project
-    |> assoc(:surveys)
-    |> Repo.get!(id)
+    project = load_project_for_change(conn, project_id)
+    survey = load_survey(project, id)
 
     unless Survey.deletable?(survey), do: raise ConflictError
 
@@ -314,21 +278,13 @@ defmodule Ask.SurveyController do
       end
     end
 
-    activity_log = fn survey -> ActivityLog.start(survey.project, conn, survey) end
-
-    project =
-      conn
-      |> load_project_for_change(project_id)
-
-    survey =
-      project
-      |> assoc(:surveys)
-      |> Repo.get!(survey_id)
+    project = load_project_for_change(conn, project_id)
+    survey = load_survey(project, survey_id)
 
     case perform_action.(survey) do
       {:ok, %{survey: survey}} ->
         Project.touch!(survey.project)
-        activity_log.(survey) |> Repo.insert!()
+        ActivityLog.start(project, conn, survey) |> Repo.insert!()
         render_survey(conn, survey)
 
       {:error, %{survey: survey}} ->
@@ -344,290 +300,19 @@ defmodule Ask.SurveyController do
     end
   end
 
-  defp render_survey(conn, survey),
-    do:
-      render(conn, "show.json",
-        survey:
-          survey
-          |> Survey.with_links(user_level(survey.project_id, current_user(conn).id))
-      )
-
-  def simulate_questionanire(conn, %{"project_id" => project_id, "questionnaire_id" => questionnaire_id, "phone_number" => phone_number, "mode" => mode, "channel_id" => channel_id}) do
-    project = conn
-    |> load_project_for_change(project_id)
-
-    questionnaire = Repo.one!(from q in Questionnaire,
-      where: q.project_id == ^project.id,
-      where: q.id == ^questionnaire_id)
-
-    channel = Repo.get!(Channel, channel_id)
-
-    survey = %Survey{
-      simulation: true,
-      project_id: project.id,
-      name: questionnaire.name,
-      mode: [[mode]],
-      state: "ready",
-      cutoff: 1,
-      schedule: Ask.Schedule.always()}
-    |> Ecto.Changeset.change
-    |> Repo.insert!
-
-    respondent_group = %RespondentGroup{
-      survey_id: survey.id,
-      name: "default",
-      sample: [phone_number],
-      respondents_count: 1}
-    |> Ecto.Changeset.change
-    |> Repo.insert!
-
-    %Ask.SurveyQuestionnaire{
-      survey_id: survey.id,
-      questionnaire_id: String.to_integer(questionnaire_id)}
-    |> Ecto.Changeset.change
-    |> Repo.insert!
-
-    %Ask.RespondentGroupChannel{
-      respondent_group_id: respondent_group.id,
-      channel_id: channel.id,
-      mode: mode}
-    |> Ecto.Changeset.change
-    |> Repo.insert!
-
-    %Respondent{
-      survey_id: survey.id,
-      respondent_group_id: respondent_group.id,
-      phone_number: phone_number,
-      sanitized_phone_number: phone_number,
-      canonical_phone_number: phone_number,
-      hashed_number: phone_number}
-    |> Ecto.Changeset.change
-    |> Repo.insert!
-
-    conn
-    |> launch(%{"project_id" => survey.project_id, "survey_id" => survey.id})
-  end
-
-  def simulation_status(conn, %{"project_id" => project_id, "survey_id" => survey_id}) do
-    survey = conn
-    |> load_project(project_id)
-    |> assoc(:surveys)
-    |> where([s], s.simulation)
-    |> Repo.get!(survey_id)
-
-    # The simulation has only one respondent
-    respondent = Repo.one!(from r in Respondent,
-      where: r.survey_id == ^survey.id)
-
-    responses = respondent
-    |> assoc(:responses)
-    |> Repo.all
-    |> Enum.map(fn response ->
-      {response.field_name, response.value}
-    end)
-    |> Enum.into(%{})
-
-    session = respondent.session
-    session =
-      if session do
-        Session.load(session)
-      else
-        nil
-      end
-
-    {step_id, step_index} =
-      if session do
-        {
-          Session.current_step_id(session),
-          case Session.current_step_index(session) do
-            {section_index, step_index} -> [section_index, step_index]
-            index -> index
-          end
-        }
-      else
-        {nil, nil}
-      end
-
-    conn
-    |> json(%{
-      "data" => %{
-        "state" => respondent.state,
-        "disposition" => respondent.disposition,
-        "step_id" => step_id,
-        "step_index" => step_index,
-        "responses" => responses,
-      }
-    })
-  end
-
-  def stop_simulation(conn, %{"project_id" => project_id, "survey_id" => survey_id}) do
-    project = conn
-    |> load_project(project_id)
-
-    survey = project
-    |> assoc(:surveys)
-    |> where([s], s.simulation)
-    |> Repo.get!(survey_id)
-
-    questionnaire = survey
-    |> assoc(:questionnaires)
-    |> Repo.one!
-
-    Repo.delete!(survey)
-    Project.touch!(project)
-
-    conn
-    |> json(%{
-      "data" => %{
-        "questionnaire_id" => questionnaire.snapshot_of
-      }
-    })
+  defp render_survey(conn, survey) do
+    level = user_level(survey.project_id, current_user(conn).id)
+    render(conn, "show.json", survey: Survey.with_links(survey, level))
   end
 
   def config(conn, _params) do
     render(conn, "config.json", config: Survey.config_rates())
   end
 
-  def create_link(conn, %{"project_id" => project_id, "survey_id" => survey_id, "name" => target_name}) do
-
-    project = conn
+  def stop(conn, %{"project_id" => project_id, "survey_id" => survey_id}) do
+    survey = conn
     |> load_project_for_change(project_id)
-
-    survey = project
-    |> assoc(:surveys)
-    |> Repo.get!(survey_id)
-
-    {name, target} = case target_name do
-      "results" ->
-        {
-          Survey.link_name(survey, :results),
-          project_survey_respondents_results_path(conn, :results, project, survey, %{"_format" => "csv"})
-        }
-      "incentives" ->
-        authorize_admin(project, conn)
-        {
-          Survey.link_name(survey, :incentives),
-          project_survey_respondents_incentives_path(conn, :incentives, project, survey, %{"_format" => "csv"})
-        }
-      "interactions" ->
-        authorize_admin(project, conn)
-        {
-          Survey.link_name(survey, :interactions),
-          project_survey_respondents_interactions_path(conn, :interactions, project, survey, %{"_format" => "csv"})
-        }
-      "disposition_history" ->
-        {
-          Survey.link_name(survey, :disposition_history),
-          project_survey_respondents_disposition_history_path(conn, :disposition_history, project, survey, %{"_format" => "csv"})
-        }
-      _ ->
-        Logger.warn "Error when creating link #{target_name}"
-        conn
-        |> put_status(:unprocessable_entity)
-        |> send_resp(:no_content, target_name)
-    end
-
-    multi = Multi.new
-    |> Multi.run(:generate_link, fn _, _ -> ShortLink.generate_link(name, target) end)
-    |> Multi.insert(:log, ActivityLog.enable_public_link(project, conn, survey, target_name))
-    |> Repo.transaction
-
-    case multi do
-      {:ok, %{generate_link: link}} ->
-        render(conn, "link.json", link: link)
-      {:error, _, changeset, _} ->
-        Logger.warn "Error when creating link #{name}"
-        conn
-        |> put_status(:unprocessable_entity)
-        |> put_view(Ask.ChangesetView)
-        |> render("error.json", changeset: changeset)
-    end
-  end
-
-  def refresh_link(conn, %{"project_id" => project_id, "survey_id" => survey_id, "name" => target_name}) do
-    project = conn
-    |> load_project_for_change(project_id)
-
-    if target_name == "interactions" || target_name == "incentives" do
-      authorize_admin(project, conn)
-    end
-
-    survey = project
-    |> assoc(:surveys)
-    |> Repo.get!(survey_id)
-
-    link = ShortLink
-    |> Repo.get_by(name: Survey.link_name(survey, String.to_atom(target_name)))
-
-    if link do
-      multi = Multi.new
-      |> Multi.run(:regenerate, fn _, _ -> ShortLink.regenerate(link) end)
-      |> Multi.insert(:log, ActivityLog.regenerate_public_link(project, conn, survey, target_name))
-      |> Repo.transaction
-
-      case multi do
-        {:ok, %{regenerate: new_link}} ->
-          render(conn, "link.json", link: new_link)
-        {:error, _, changeset, _} ->
-          Logger.warn "Error when regenerating results link #{inspect link}"
-          conn
-          |> put_status(:unprocessable_entity)
-          |> put_view(Ask.ChangesetView)
-          |> render("error.json", changeset: changeset)
-      end
-    else
-      Logger.warn "Error when regenerating results link #{target_name}"
-      conn
-      |> put_status(:unprocessable_entity)
-      |> send_resp(:no_content, target_name)
-    end
-  end
-
-  def delete_link(conn, %{"project_id" => project_id, "survey_id" => survey_id, "name" => target_name}) do
-    project = conn
-    |> load_project_for_change(project_id)
-
-    if target_name == "interactions" || target_name == "incentives" do
-      authorize_admin(project, conn)
-    end
-
-    survey = project
-    |> assoc(:surveys)
-    |> Repo.get!(survey_id)
-
-    link = ShortLink
-    |> Repo.get_by(name: Survey.link_name(survey, String.to_atom(target_name)))
-
-    if link do
-      multi = Multi.new
-      |> Multi.delete(:delete, link)
-      |> Multi.insert(:log, ActivityLog.disable_public_link(project, conn, survey, link))
-      |> Repo.transaction
-
-      case multi do
-        {:ok, _} -> send_resp(conn, :no_content, "")
-        {:error, _, changeset, _} ->
-          Logger.warn "Error when deleting link #{inspect link}"
-          conn
-          |> put_status(:unprocessable_entity)
-          |> put_view(Ask.ChangesetView)
-          |> render("error.json", changeset: changeset)
-      end
-    else
-      conn
-      |> send_resp(:not_found, "")
-    end
-  end
-
-  def stop(conn, %{"project_id" => project_id, "survey_id" => id}) do
-    project =
-      conn
-      |> load_project_for_change(project_id)
-
-    survey =
-      project
-      |> assoc(:surveys)
-      |> Repo.get!(id)
+    |> load_survey(survey_id)
 
     case SurveyAction.stop(survey, conn) do
       {:ok, %{survey: survey, cancellers_pids: cancellers_pids}} ->
@@ -660,19 +345,13 @@ defmodule Ask.SurveyController do
   end
 
   def update_locked_status(conn, %{"project_id" => project_id, "survey_id" => survey_id, "locked" => locked}) do
-    project =
-      conn
-      |> load_project_for_owner(project_id)
+    project = load_project_for_owner(conn, project_id)
 
-    survey =
-      project
-      |> assoc(:surveys)
-      |> Repo.get!(survey_id)
-
-    survey = survey
+    survey = project
+    |> load_survey(survey_id)
     |> Repo.preload([:quota_buckets])
     |> Repo.preload(:questionnaires)
-    |> Survey.with_links(user_level(survey.project_id, current_user(conn).id))
+    |> Survey.with_links(user_level(project_id, current_user(conn).id))
 
     case survey.state do
       "running" ->
@@ -710,37 +389,9 @@ defmodule Ask.SurveyController do
     end
   end
 
-  def simulation_initial_state(conn, %{"project_id" => project_id, "survey_id" => survey_id, "mode" => mode}) do
-    survey = load_project(conn, project_id)
+  defp load_survey(project, survey_id) do
+    project
     |> assoc(:surveys)
-    |> where([s], s.simulation)
     |> Repo.get!(survey_id)
-
-    render_initial_state(conn, survey.id, mode)
-  end
-
-  defp render_initial_state(conn, survey_id, "mobileweb" = _mode) do
-    # The simulation has only one respondent
-    respondent = Repo.one!(from r in Respondent,
-    where: r.survey_id == ^survey_id)
-
-    response = if (respondent.session) do
-      session = Session.load_respondent_session(respondent, true)
-
-      json(conn, %{
-        "data" => %{
-          "mobile_contact_messages" => Session.mobile_contact_message(session)
-        }
-      })
-    else
-      conn
-      |> send_resp(:not_found, "")
-    end
-
-    response
-  end
-
-  defp render_initial_state(conn, _survey_id, _mode) do
-    json(conn, %{"data" => %{}})
   end
 end
